+++ tests/open-animated.test.js
@@ -1,0 +1,8 @@
+import assert from 'assert';
+import GUI from '..';
+
+export default () => {
+
+	assert.doesNotThrow( () => new GUI().openAnimated() );
+
+};
